% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_calendario.R
\name{constroi_calendario}
\alias{constroi_calendario}
\title{constroi calendario das disciplinas}
\usage{
constroi_calendario(dataset_horario)
}
\arguments{
\item{dataset_horario}{dataset que contem as disciplinas construido pela funcao pega_info}
}
\description{
constroi calendario das disciplinas
}
