% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_info.R
\name{pega_discp}
\alias{pega_discp}
\title{pega disciplinas do sigaa}
\usage{
pega_discp(area = 3, campus)
}
\arguments{
\item{area}{da graduacao,}

\item{campus}{campus em que deve buscar as disciplinas}
}
\description{
pega disciplinas do sigaa
}
